+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export { default as APIPage } from "./APIPage";
 export { default as Home } from "./Home";
 export { default as Login } from "./Login";
-export { default as Signup } from "./Signup";
+export { default as Signup } from "./SignUp/Signup";
